cleanup + cxxbuild system

+++ src/core/mesh.cpp
@@ -1,6 +1,6 @@
 
 #include <mesh.h>
 
-Mesh::Mesh(GLuint vaoid, size_t vcount) : vao_id(vaoid), vert_count(vcount) {;;}
-GLuint Mesh::vao(){return vao_id;}
+Mesh::Mesh(GLuint vaoid, size_t vcount) : vao_id(vaoid), vert_count((GLuint)vcount) {;;}
+GLuint Mesh::vao(){return vao_id;} 
 GLuint Mesh::vcount(){return (GLuint)vert_count;}
